requirements.txt
fastapi 
uvicorn
sqlalchemy 
psycopg2-binary 
python-dotenv 
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
pymysql
PyJWT
fastapi-mail
python-multipart
alembic
mysqlclient